Console/requirements.txt
argon2-cffi
psycopg2-binary